examples/streamlit_demo/requirements.txt
streamlit==1.28.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.27.0
manul-tracer